Move TestConfig class to under spec/lib/ to be reusable
